refactor(inception): Fixed errors and refactored codebase to improve maintainability
