+++ pyproject.toml
@@ -10,11 +10,11 @@ packages = [{include = "inception"}]
 python = ">=3.9,<3.13"
 fastapi = "^0.104.1"
 uvicorn = {extras = ["standard"], version = "^0.24.0"}
+torch = "^2.5.0"
 sentence-transformers = "^3.0.0"
 pydantic = "^2.4.2"
 pydantic-settings = "^2.0.3"
 nltk = "^3.8.1"
-torch = "2.1.0+cu121"  # Match CUDA 12.1
 numpy = "^1.26.1"
 gunicorn = "^21.2.0"
 prometheus-client = "^0.19.0"
@@ -33,3 +33,6 @@ pytest-asyncio = "^0.21.1"  # For async test support
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.group.gpu.dependencies]
+torch = {version = "^2.5.0+cu121", source = "pypi"}
